+++ packages/fastboard-ui/src/typings.ts
@@ -15,3 +15,18 @@ export type GenericIcon<K extends string, E extends string = IconType> = {
 };
 
 export type I18nData<T extends string> = Record<Language, Record<T, string>>;
+
+export interface FastboardUIConfig {
+  toolbar?: {
+    enable?: boolean;
+  };
+  redo_undo?: {
+    enable?: boolean;
+  };
+  zoom_control?: {
+    enable?: boolean;
+  };
+  page_control?: {
+    enable?: boolean;
+  };
+}