+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - Re-export [`@netless/app-slide`](https://github.com/netless-io/netless-app/tree/master/packages/app-slide).
+- Rename `@netless/app-media-player` to `@netless/app-plyr`.
+- Update toolbar handler styles.
+- Added `config` field to the `Fastboard` component to hide controls.
 
 ## 0.3.1